Split file explorer git status coverage
